requirements.txt
Cython
websocket-client
plyer